absorb into util
