+++ scripts/yodaPlotNPCorr.py
@@ -12,32 +12,11 @@ import numpy as np
 import yoda
 
 from fit import scipy_fit as fit
-from util import NumpyEncoder, json_numpy_obj_hook
+from util import NumpyEncoder, valid_yoda_file
 
 
 COLORS = ["#e41a1c", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33", "#a65628", "#f781bf", "#999999"]
 XTICKS = [1, 2, 5, 10, 20, 50, 100, 200, 500, 1000, 2000, 5000]
-
-
-def valid_yoda_file(param):
-    """Helper function which checks for validity (YODA extension and existence) of the provided input files
-
-    Args:
-        param (AnyStr): File to check
-
-    Raises:
-        argparse.ArgumentTypeError: Wrong file extension
-        IOError: No such file
-
-    Returns:
-        AnyStr@abspath: 
-    """
-    base, ext = os.path.splitext(param)
-    if ext.lower() not in ('.yoda'):
-        raise argparse.ArgumentTypeError('File must have a yoda extension')
-    if not os.path.exists(param):
-        raise IOError('{}: No such file'.format(param))
-    return os.path.abspath(param)
 
 
 parser = argparse.ArgumentParser(
